A: Split Requests
