+++ API/Database/Requests/ClientRequests.go
@@ -2,7 +2,6 @@ package Requests
 
 import (
 	"API/Database"
-	"API/Database/Common"
 	"API/Models"
 	"fmt"
 )
@@ -21,24 +20,4 @@ func GetClientProfileInfo(pMembershipNumber int) []Models.Client {
 
 	return client
 
-}
-
-func GetReservedSessions(pMembershipNumber int) Models.Schedule {
-
-	query := fmt.Sprintf(`EXEC SP_getBookings %d;`, pMembershipNumber)
-
-	resultSet, err := Database.ReadTransaction(query)
-
-	if err != nil {
-		return Models.Schedule{}
-	}
-
-	schedule := ParseSchedule(resultSet)
-
-	return schedule
-}
-
-func RegisterClientUser(pUsername string, pPassword string, pMembershipNumber int) Common.VoidOperationResult {
-	query := fmt.Sprintf(`EXEC SP_RegisterClientUser '%s', '%s', %d;`, pUsername, pPassword, pMembershipNumber)
-	return VoidRequest(query)
 }
