+++ API/Database/Requests/UserRequests.go
@@ -2,11 +2,12 @@ package Requests
 
 import (
 	"API/Database"
+	"API/Database/Common"
 	"API/Models"
 	"fmt"
 )
 
-func GetUser(pUsername string) (Models.Login, bool) {
+func GetLogin(pUsername string) (Models.Login, bool) {
 
 	query := fmt.Sprintf(`EXEC SP_GetUserByUsername '%s';`, pUsername)
 
@@ -18,4 +19,19 @@ func GetUser(pUsername string) (Models.Login, bool) {
 
 	user := ParseLoginResponse(resultSet)
 	return user, true
+}
+
+func DeactivateAccount(pUsername string, pUserTypeID int) Common.VoidOperationResult {
+	query := fmt.Sprintf(`EXEC SP_DeactivateAccount '%s', %d;`, pUsername, pUserTypeID)
+	return VoidRequest(query)
+}
+
+func UpdateUserDetails(pOldUsername string, pNewUsername string, pUserTypeID int) Common.VoidOperationResult {
+	query := fmt.Sprintf(`EXEC SP_UpdateUserDetails '%s', '%s', %d;`, pOldUsername, pNewUsername, pUserTypeID)
+	return VoidRequest(query)
+}
+
+func RegisterClientUser(pUsername string, pPassword string, pMembershipNumber int) Common.VoidOperationResult {
+	query := fmt.Sprintf(`EXEC SP_RegisterClientUser '%s', '%s', %d;`, pUsername, pPassword, pMembershipNumber)
+	return VoidRequest(query)
 }
